proposal: Arbitrary broker resource type
This proposal aims to find a way to use arbitrary resources as brokers,
thus allowing interaction between Akri and external projects like for
example Shifu.
As these projects uses CustomResources to depict their workload, we
need to be able to create arbitrary resources.

This is quite a major change to the Configuration model.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>